Restructured the package to inclued another nest of dftintegrate following the example of fortpy.  Also added packages, scripts, and package_data to setup.py in hopes to make the package importable once pip installed.
