+++ dftintegrate/fourierfit/vaspdata.py
